+++ pyproject.toml
@@ -48,6 +48,7 @@ pytest-cov = ">=3"
 pytest-xdist = ">=3"
 pytest-asyncio = ">=0.20"
 pytest-mock = ">=3"
+hypothesis = ">=6.80"
 mypy = ">=1"
 mkdocs = ">=1.4"
 mkdocs-material = ">=9"
