+++ reprisal/hooks/types.py
@@ -1,7 +1,8 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
 from typing import Callable, Coroutine, Generic, TypeVar
+
+from pydantic.generics import GenericModel
 
 T = TypeVar("T")
 
@@ -11,6 +12,5 @@ Setup = Callable[[], Coroutine[None, None, None]]
 Deps = tuple[object, ...] | None
 
 
-@dataclass(slots=True)
-class Ref(Generic[T]):
+class Ref(GenericModel, Generic[T]):
     current: T
